Move fn-execution org.apache.beam.harness.channel
'harness' suggests the SDK harness package, which is not correct, as
this is used by both runners and the SDK harness.
