+++ runners/java-fn-execution/src/main/java/org/apache/beam/runners/fnexecution/ServerFactory.java
@@ -28,7 +28,7 @@ import java.io.IOException;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
-import org.apache.beam.harness.channel.SocketAddressFactory;
+import org.apache.beam.sdk.fn.channel.SocketAddressFactory;
 import org.apache.beam.model.pipeline.v1.Endpoints;
 
 /**
