+++ sdks/java/fn-execution/src/main/java/org/apache/beam/sdk/fn/channel/SocketAddressFactory.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.beam.harness.channel;
+package org.apache.beam.sdk.fn.channel;
 
 import static com.google.common.base.Preconditions.checkArgument;
