+++ sdks/java/fn-execution/src/main/java/org/apache/beam/sdk/fn/channel/package-info.java
@@ -19,4 +19,4 @@
 /**
  * gRPC channel management.
  */
-package org.apache.beam.harness.channel;
+package org.apache.beam.sdk.fn.channel;
